@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version:6.2
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "cacherun",
     platforms: [
-        .macOS(.v10_15)
+        .macOS(.v26)
     ],
     products: [
         .executable(name: "cacherun", targets: ["cacherun"]),
@@ -16,21 +16,20 @@ let package = Package(
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
         .package(url: "https://github.com/apple/swift-argument-parser", .upToNextMajor(from: "1.2.0")),
-        .package(url: "https://github.com/apple/swift-tools-support-core", .upToNextMajor(from: "0.6.0"))
+        .package(url: "https://github.com/apple/swift-system", from: "1.6.1"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages which this package depends on.
-        .target(
+        .executableTarget(
             name: "cacherun",
             dependencies: [
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
-                .product(name: "SwiftToolsSupport-auto", package: "swift-tools-support-core"),
                 .target(name: "CacheExecutor")
             ]),
         .target(name: "CacheExecutor",
                 dependencies: [
-                    .product(name: "SwiftToolsSupport-auto", package: "swift-tools-support-core")
+                    .product(name: "SystemPackage", package: "swift-system")
                 ]),
         .testTarget(
             name: "cacherunTests",
