@@ -7,7 +7,7 @@
 
 import Foundation
 import XCTest
-import TSCBasic
+import SystemPackage
 @testable import CacheExecutor
 
 final class cachExecutorTests: XCTestCase {
@@ -19,16 +19,9 @@ final class cachExecutorTests: XCTestCase {
         return try! OutputCachingExecutor.Utility.sha256Hash(for: commandPlusArgs.joined(separator: " "))
     }
 
-    lazy var runDirectory: AbsolutePath = {
+    lazy var runDirectory: FilePath = {
         return OutputCachingExecutor.Utility.locateRunDirectory()
     }()
-
-    override func tearDown() {
-        let pidFile = AbsolutePath(runDirectory, "\(commandBaseFilename).pid")
-        let cacheFile = AbsolutePath(runDirectory, "\(commandBaseFilename).data")
-        try! localFileSystem.removeFileTree(pidFile)
-        try! localFileSystem.removeFileTree(cacheFile)
-    }
 
     func testCachedExecution() {
         let outputStream = MockOutputByteStream()
@@ -43,8 +36,8 @@ final class cachExecutorTests: XCTestCase {
         outputStream.reset()
 
         // inject a different date into the cache file, and expect the executor to return the new cached data
-        let cacheFile = AbsolutePath(runDirectory, "\(commandBaseFilename).data")
-        try! "Thu Jan  2 00:00:00 UTC 1970\n".write(toFile: cacheFile.pathString, atomically: true, encoding: .utf8)
+        let cacheFile = runDirectory.appending("\(commandBaseFilename).data")
+        try! "Thu Jan  2 00:00:00 UTC 1970\n".write(toFile: cacheFile.string, atomically: true, encoding: .utf8)
         if case let .failure(executorError) = executor.runCachedCommand() {
             XCTFail("error running executor \(executorError)")
         }
@@ -57,6 +50,12 @@ final class cachExecutorTests: XCTestCase {
             XCTFail("error running executor \(executorError)")
         }
         XCTAssertEqual(outputStream.trimmedString, "Thu Jan  1 00:00:00 UTC 1970")
+
+        let pidFile = runDirectory.appending("\(commandBaseFilename).pid")
+        let cmdFile = runDirectory.appending("\(commandBaseFilename).cmd")
+        XCTAssertFalse(pidFile.isFile())
+        try! cmdFile.delete()
+        try! cacheFile.delete()
     }
 
     func testUnknownCommandError() {
@@ -65,58 +64,94 @@ final class cachExecutorTests: XCTestCase {
         let executor = OutputCachingExecutor(cacheTime: 1, userCommand: ["/fake/command"], outputStream: outputStream, errorStream: errorStream)
         switch executor.runCachedCommand() {
         case .failure(let error):
-            guard case CacheExecutorError.systemError(_) = error else {
-                return XCTFail("expected a system error")
+            guard case CacheExecutorError.badCommand(_) = error else {
+                return XCTFail("expected a bad command error")
             }
         default:
             XCTFail("expected a system error")
         }
+        let pidFile = runDirectory.appending("\(commandBaseFilename).pid")
+        let cacheFile = runDirectory.appending("\(commandBaseFilename).data")
+        let cmdFile = runDirectory.appending("\(commandBaseFilename).cmd")
+        XCTAssertFalse(pidFile.isFile())
+        XCTAssertFalse(cacheFile.isFile())
+        XCTAssertFalse(cmdFile.isFile())
     }
 
     func testCommandError() {
         let outputStream = MockOutputByteStream()
         let errorStream = MockOutputByteStream()
-        let executor = OutputCachingExecutor(cacheTime: 1, userCommand: ["/bin/ls", "i-dont-exist-in-the-fs"], outputStream: outputStream, errorStream: errorStream)
+        let commandPlusArgs = ["/bin/ls", "i-dont-exist-in-the-fs"]
+        let executor = OutputCachingExecutor(cacheTime: 1, userCommand: commandPlusArgs, outputStream: outputStream, errorStream: errorStream)
         if case let .failure(executorError) = executor.runCachedCommand() {
             XCTFail("error running executor \(executorError)")
         }
         XCTAssertEqual(errorStream.trimmedString, "ls: i-dont-exist-in-the-fs: No such file or directory")
+
+        let commandBaseFilename = try! OutputCachingExecutor.Utility.sha256Hash(for: commandPlusArgs.joined(separator: " "))
+        let pidFile = runDirectory.appending("\(commandBaseFilename).pid")
+        let cacheFile = runDirectory.appending("\(commandBaseFilename).data")
+        let cmdFile = runDirectory.appending("\(commandBaseFilename).cmd")
+        XCTAssertFalse(pidFile.isFile())
+        try! cmdFile.delete()
+        try! cacheFile.delete()
     }
 
 }
-
-public final class MockOutputByteStream: OutputByteStream {
-
+public final class MockOutputByteStream: Utf8TextOutputStream {
     /// Contents of the stream.
-    private var contents = [UInt8]()
-
-    /// The contents of the output stream.
-    ///
-    /// - Note: This implicitly flushes the stream.
-    public var bytes: ByteString {
-        return ByteString(contents)
-    }
+    private let contents = LockIsolated(String())
 
     public var trimmedString: String {
-        return bytes.cString.trimmingCharacters(in: .whitespacesAndNewlines)
+        contents.update {
+            $0.trimmingCharacters(in: .whitespacesAndNewlines)
+        }
     }
 
     /// The current offset within the output stream.
     public final var position: Int {
-        return contents.count
+        return contents.get().count
     }
 
-    public func write(_ byte: UInt8) {
-        contents.append(byte)
+    public func write(_ string: String) {
+        contents.update {
+            $0.appending(string)
+        }
     }
 
     public func reset() {
-        contents = [UInt8]()
+        contents.set("")
     }
 
     public func flush() { }
+}
 
-    public func write<C>(_ bytes: C) where C : Collection, C.Element == UInt8 {
-        contents.append(contentsOf: bytes)
+final class LockIsolated<T>: @unchecked Sendable {
+    private let lock = NSLock()
+    private var value: T
+
+    init(_ value: T) {
+        self.value = value
     }
+
+    func get() -> T {
+        lock.lock()
+        defer { lock.unlock() }
+        return value
+    }
+
+    func set(_ newValue: T) {
+        lock.lock()
+        defer { lock.unlock() }
+        value = newValue
+    }
+
+    @discardableResult
+    func update(_ closure: (T) -> T) -> T {
+        lock.lock()
+        defer { lock.unlock() }
+        value = closure(value)
+        return value
+    }
+
 }
